fix(controller): 요청 데이터에 대해 null 값 허용

+++ src/main/java/roomescape/service/ThemeService.java
@@ -31,7 +31,10 @@ public class ThemeService {
                 .orElseThrow(() -> new BadRequestException("[ERROR] 데이터가 저장되지 않습니다."));
     }
 
-    public void deleteTheme(long id) {
+    public void deleteTheme(Long id) {
+        if (id == null) {
+            throw new BadRequestException("[ERROR] id에 null이 입력될 수 없습니다.");
+        }
         themeRepository.deleteThemeById(id);
     }
